File truncate is now synchronous
This is mostly to ensure that both couch_file and the writer process
are keeping track of the exact same Eof value. Without this, write
operations followed by a truncate, which is then immediately followed
by other writes may cause an Eof inconsistency between the 2 processes,
especially if the writes come from different processes.

In practice this doesn't happen, since in the only place where we use truncate
we do it as the first file operation (without any previous write) and it is
not immediately followed by write operations.