Replication manager: allow edition of replication documents
Replication documents that are not in the triggered state can now
be edited by users. This is to make it simpler for Futon users
to restart replications - they can just edit the document, remove
its state field, and save it, instead of recreating it or adding
a similar document but with a different _id.

Change-Id: I40c2cdccb882a4e49f08e0932ed7565738a33bb3
git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@1125319 13f79535-47bb-0310-9956-ffa450edef68
Reviewed-on: http://review.membase.org/6399
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Dustin Sallings <dustin@spy.net>
Reviewed-by: Dustin Sallings <dustin@spy.net>