Add CoffeeScript (http://coffeescript.org/) support to the JS view server.
This allows you to write all query server functions in CoffeeScript.

The implementation re-uses the JS query server and adds an additional,
optional compilation step. To make the implementation easier for now,
a new main-coffee.js file is generated that includes the CoffeeScript
compiler, but I think we can do away with that in a future optimisation
and just use a single main.js that can handle both.

The patch includes basic tests for the JS test suite that I tested
in Firefox and Chrome. I'm no expert in CoffeeScript, so I don't
know if more complex script will break the implementation. `make
distcheck` passes.

Futon picks up the "coffeescript" design document lanugage as expected
and view definition interaction works as usual.

The implementation does some sting maniulation on top of the
CoffeeScript compile step that probably can be done more nicely.

List functions don't deal well with the implicit return generated by
CoffeeScript, but J Chris says, this will be easy to fix.

The patch adds CoffeeScript 1.1.1 which is MIT licensed and updates
the NOTICE, license.skip and .gitignore files as well as the build
system accordingly.

Change-Id: I85d992a6c8fdad33c7ea6f1e10eddae49b915fec
Reviewed-on: http://review.couchbase.org/6640
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>