Update seq format/datatype agnostic replicator
Now the replicator no longer expects the update
seqs to be numbers in order to determine between
N _changes rows which came earlier. When it receives
a _changes row, it tags it with a timestamp (a counter
in fact) and uses that timestamp to determine the
earliest row in a group.

This is necessary to ensure that pull replications
from BigCouch work (just like they did in the old
replicator).

Change-Id: I586349c3d067bd2a111d376f12c5c7550171ca31
Reviewed-on: http://review.couchbase.org/7820
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>