Refactor/simplify replicator
For each worker (doc copier) process there used to be
a rev finder process. A rev finder dequeues from the
_changes rows queue, finds which document IDs and revisions
are missing in the target and queues the IDs and revisions
of those that are missing into a second queue. Finally
worker processes dequeue up to N elements from this queue.

This model was reduced to have the workers dequeing
directly from the _changes rows queue and find which document
IDs and revisions are missing in the target. This makes the
model simpler without suffering any performance penalty - by
the contrary due to less process communication it seems that
replications are making more steady progress.
The processing of which revisions are missing in the target
was also significantly simplified.

Change-Id: I36d0573ce18156e5c6811e0da341a8e5452d420e
Reviewed-on: http://review.couchbase.org/7858
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>