Gracefully hand invalid compaction configuration
If the compaction configuration for a database is invalid, don't
crash the daemon - just catch the error and log an helpful error
message.

This could make the daemon be restarted 10 times in less then an
hour by the CouchDB main supervisor - after the 10th attempt, Couch
would need to be restarted (the supervisor's configuration allows for
a maximum restart frequency of 10 per hour).

Change-Id: Id7c9d45d841142bbc003019432e760d0dc253eb3
Reviewed-on: http://review.couchbase.org/8055
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>