Don't spawn new compaction process on retry
After the compaction process finishes, the updater checks
if there were updates done immediately after the compaction
process finished. If there were, it retries and spawns a new
process to retry the compaction.

This change makes sure the same process is used for the retries.
This is necessary for the compaction daemon to take into account
the time taken by the retries when there's an allowed period
configured.

This is already done for view compaction but not for database
compaction (it was forgotten before).

Change-Id: I3f52444ce82ba4135b85259893aef63e4728a150
Reviewed-on: http://review.couchbase.org/8058
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>