Add a "Verify Installation" page to Futon.
The "verify install" procedure replaces	the full JS test suite for users
to see if they installation is correct. The tests run check for CouchDB's
basic feature set and should notify users of the most common installation
issues.

The full test suite is demoted to be "for Developers" in the sidebar, so
that core CouchDB developers can keep using this for development and
release testing, but end-users won't have to deal with the various
browser and runtime issues on different hardware.

git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@1147832 13f79535-47bb-0310-9956-ffa450edef68

Conflicts:

	share/www/_sidebar.html