MB-6591: Use raw collation for _all_docs query params
_all_docs are always sorted in raw order, not Unicode. Hence
the query parameters (startkey and endkey) also need to
take this raw collation into account. The query validation checks
whether startkey is smaller than endkey, hence we need to take
the collation type (Unicode or raw) into account.

An example for raw collation is:

    Foo < bar < foo < zoo

With Unicode collation it would be:

    bar < foo < Foo < zoo

This patch is based on a patch for Apache CouchDB [1].

I've kept the original commit message below.

[1] https://github.com/apache/couchdb/commit/46d2dc647f786fa1920f365af89ce4d0070ef280

Fix empty range check for raw collation.

The check for empty ranges was not taking into account the
view option for raw collation. This fixes that by passing
the couch_btree:less/2 function into the check.

Patch by: Jason Smith
Re: COUCHDB-1228 4/4

Change-Id: I4569683ff0253221f2f265e287b7a52d2ad32ac1
git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@1156509 13f79535-47bb-0310-9956-ffa450edef68
Reviewed-on: http://review.couchbase.org/21359
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>