Fix process deadlock in the os process pool
It's possible to reach a state where we get the
os pool full (os_process_limit processes loaded) and
it never gets updated when clients which requested a
proc die (or don't explicitly release the os proc when
they're done with it). This prevents other clients
from getting an OS proc, being blocked forever by the
pool.

Change-Id: I6b2e61d8dc9bd70acb190db683b10c24cf335f7e
Reviewed-on: http://review.couchbase.org/9039
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>