Synchronize view execution with design document replication.
Due to lags in design document replication it was possible to get view
results corresponding to different design document versions from
different nodes. Now view merger ensures that all the nodes have the
same version of design document.

'view_merge' record passed to couch_view_merger:query_view now has
additional field 'ddoc_revision'. It can be 'nil' (default) which
means that the caller does not care about ddoc consistency, 'auto'
which means that query_view should peek a revision on its own and
return the result whenever a consistent state is achieved or it can be
certain design document revision which means that query_view should
try to use this revision but in case of inconsistency an exception
will be thrown. A _view_merge handler is changed accordingly so that
ddoc_revision can be specified in merge spec.

Limitations:

  1. When ddoc_revision is specified it is implicitly assumed that all
  the view specs refer to the same design document.

  2. Remote databases are not supported so they must be avoided in
  specs. Though the same result can be obtained by merged spec
  containing one local database.

  3. Nested 'ddoc_revision' specs are not allowed.

  4. It's impossible to determine reliably whether a connection is
  still alive and thus impossible to perform indefinite wait for
  synchronization (otherwise in unfortunate situation we could end up
  with plenty of hanging processes). Because of that only a fixed
  amount of retries are performed and an {error, revision_sync_failed}
  exception is thrown if synced state has not been reached.

Change-Id: Ic64e6f19403df253cb4df23f5c2c437cc2835d26
Reviewed-on: http://review.couchbase.org/9573
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>