Use a local UUID, not hostname, in replication ID
In a mobile app, the hostname and port are not at all
constant between launches, so they're not good to use for
generating replication IDs. This is causing replication to
start over from scratch on every launch, making it very
slow.

Filipe gave me (Jens) this patch, which creates a local
UUID for the server, stores it in the config, and uses it
instead of the hostname+port in the replication ID.

Change-Id: I2bc37d3ffb2b5181cbe46decea6399f499a66b2d
Reviewed-on: http://review.couchbase.org/9575
Reviewed-by: Chris Anderson <jchris@couchbase.com>
Tested-by: Jens Alfke <jens@couchbase.com>