Do not suppress old revisions from _changes
Requests to '_changes?style=all_docs' with a 'since' parameter would
suppress revisions of a document generated before the start of the
feed.  The assumption may have been that those revisions were seen by
the client in a previous request to _changes, but this is not
necessarily true if the client is resuming the feed after a partial
download.

The safe thing to do is to include all leaf revisions for each document
in the feed.

Thanks Bob Dionne for the test code.

Closes COUCHDB-1256

Change-Id: Ife96f7c299fab8b2d151fab28f55cc4b39d22812
git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@1161697 13f79535-47bb-0310-9956-ffa450edef68
Reviewed-on: http://review.couchbase.org/9657
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>