View servers can add dependencies to their view definitions
This will be necessary for Objective-C views, since the view definition
simply contains the class and method names, which remain the same as the
code changes. The ObjC view server can therefore ask the class for a
version identifier, and append that to be hashed along with the view.

Conflicts:

	src/couchdb/couch_view_server_objc.erl

Change-Id: Ia997d298022909a3c0cbaf0d23f6de38c0ac7a91
Reviewed-on: http://review.couchbase.org/10022
Reviewed-by: Chris Anderson <jchris@couchbase.com>
Reviewed-by: Jonathon Mah <me@jonathonmah.com>
Tested-by: Jens Alfke <jens@couchbase.com>