Add support for set views to couch_httpd_view_merger.
All the set views are located in the "sets" property of "views". Each
spec must specify view name in the same way as for simple specs and
the list of partitions. Though it's implicitly assumed that design
document database is always named 'master' for set views.

Example of merge spec:

      { "views": {
          "sets" : {
            "default" : {
              "view" : "test/map",
              "partitions" : [0,1,2,3,4,5,6,7]
            }
          }
         }
      }

Change-Id: Ib8bf403e0c16827b7579b4de96cf90c7d7223c9d
Reviewed-on: http://review.couchbase.org/9788
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>