Replicator: update task status more often
Instead of having workers report their stats only when they
finish a batch, allow them to report them more often so that
the replication entry in _active_tasks is updated more often.
This avoids giving the wrong idea to applications that the
replicator is stuck when workers are processing heavy batches.

Change-Id: I05c83c1ceab9a81ac1a3287203408aaa1f7e3400
Reviewed-on: http://review.couchbase.org/9916
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>