Ensure bitmasks are not lost
If the updater is running and the bitmasks are updated,
make sure that when the updater finishes we don't lose any
of those bitmask updates.
Scenarios:

1) The updater is running and partitions are toggled from passive
to active - we set the new group's (group received from the updater)
bitmask to our current bitmasks (active and passive)

2) The updater is running and partitions are toggled from active to
passive - we stop immediately the updater, update the new group's
bitmasks and restart the updater

3) The updater is running and partitions are marked for cleanup - we
stop immediately the updater, update the new group's bitmasks and
restart the updater

If the updater is stopped while still indexing the active partitions,
clients who requested the group with the stale=false option will not
be unblocked. They will only be unblocked as soon as all active
paritions are processed by the updater.

Change-Id: Ic5971e672a7389678468f3e23b8ea39925820b23
Reviewed-on: http://review.couchbase.org/10145
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>