Fix view group index file descriptor leak
If a view group is opened and later the corresponding
design document is updated in a such a way that it
produces a new view group signature (for example, by
changing a map function), a new view group will be
spawned. Requests to /database/_view_cleanup will not
effectively delete the old index file because the old view
group is still alive and holding a reference counter
associated to that old index file (unless the server is
restarted).

COUCHDB-1309

Change-Id: I7e0fa6e8d669713b11b36c7674e853e680539570
Reviewed-on: http://review.couchbase.org/10165
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>