Do not trigger set view update when not necessary in view merger.
Since we do not use view group when get_{map,reduce}_view returns
non-empty list of missing partitions we do not need to trigger set
view update. So we always get set view group with stale=ok first and
only if it's needed we get it again with user-supplied stale option.

Change-Id: I942f2783d81320fb08ba16a7629e91e3397c74a8
Reviewed-on: http://review.couchbase.org/10393
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>