Avoid confusing log messages when replication completes
When a replication finishes, the replicator manager (database)
will call couch_replicator:cancel_replication/1 with the purpose
of ensuring the child spec is removed from the supervisor.
This function however is used as well when users explicitly ask
to cancel a replication, and logs messages which don't make sense
when invoked by the replication manager (added by COUCHDB-1271).
This change also adds a comment about this behaviour.

Change-Id: I08cec6cdaad216f242c689161f636d2b40220f19
Reviewed-on: http://review.couchbase.org/10503
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>