More efficient set view indexing
This change makes the communication between the Erlang VM and
an external view server (couchjs for e.g.) more efficient since
there's now a process which keeps writing to the port (pipe) and
another one which keeps reading from it. This minimizes the
amount of time each endpoint spends blocked on the port read
operation.

Change-Id: Iae5622650dc75e63f7f097a103c3ddf724e58c5f
Reviewed-on: http://review.couchbase.org/10622
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>