Update handling of set view group shutdown
If a set database (partition) is deleted, shutdown the
set view group only if that partition is marked as active
or passive in the set view group's state.
This prevents unnecessary shutdowns and index file deletes.

Also fixed a badmatch when doing a lookup in the ets table
couch_sig_to_setview_pid because the keys for this table have
the shape {SetName, Sig} and not {DDocId, Sig}.

Change-Id: Ie39f40816793ec65bf0aabff0569da6be240f889
Reviewed-on: http://review.couchbase.org/11799
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>