Fix test view_merge.py
It should verify the length of the rows array and not rely
on the total_rows attribute. The later is always the number
of rows in the view, not necessarily the number of rows
returned in the response (length(rows) != total_rows).
If the view response returned 0 rows, due to an error, the
test thought we got the right anwser.

Change-Id: I29f4b8adc984c15f259fe7dd3af202737ea98655
Reviewed-on: http://review.couchbase.org/12232
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>