Don't stop set view updater unnecessarily
Only stop it once it transitions to the update_passive
state if there are cleanup waiters. Otherwise let it run
to continue updating the passive partitions.

This could lead to a case where a continuous burst of view
queries with ?stale=false came and was constantly stopping the
updater once it got to the updating_passive state, preventing
replica partitions from ever being transferred from the replica
to the main index.

Change-Id: If2c0dce032954cffb7bd308e18a73d37f061a21b
Reviewed-on: http://review.couchbase.org/12288
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>