Propagate iolist read errors to callers
Instead of having the file reader process crash with
a badmatch after a file:pread/3, send the result to
to the caller, which will in turn get a badmatch error.
The advantage of this is to get a much more meaningful
stack trace (relative to the client) to better help
finding programmer's errors (for example a previously
missing file flush call).

Change-Id: Id3ac0a24b0ca7c3916c1d2662cb154ce8b4a0240
Reviewed-on: http://review.couchbase.org/12407
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>