Kill compaction couch_file on compactor kill/restart
If the compactor was killed, the compaction file's couch_file was
not being killed because it was linked to the view group but not
linked to the compactor process.
On compaction restart, a new couch_file for the compaction file
was being spawn and this implied that we could have 2 couch_file
processes writing (and truncating) to the same file.

This relates to MB-4650 and MB-4644

Change-Id: Idfefb29e7039235277df4638ecda75e950603849
Reviewed-on: http://review.couchbase.org/12487
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>