Sync shutdown of compactor and updater processes
The sync shutdown avoids race conditions that happen when
compaction finishes and there is an updater running. In this
case we need to shutdown the updater before doing the compact
file switch because if not, the updater's couch_file might
open the compacted file just after it was renamed, causing
all sort of badmatch errors when attempting to read from it.

This relates to MB-4650 and MB-4644

Change-Id: I19450324ffd11b646026e64efe69e6504a1f9799
Reviewed-on: http://review.couchbase.org/12491
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>