Fix test compaction_transitions.py
After recent change such as the synchronous shutdown of the
set view compactor process, this test was failing often.
This change makes it pass and is a more reliable aproach to
find out if the compaction was running when the partition
states were updated - the compactor process is restarted
when such request is made and therefore there's a tiny
window of time where there's no task in _active_tasks.

Change-Id: Icffe100c1a26e3bd0b746183b7b7338089308ace
Reviewed-on: http://review.couchbase.org/12510
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>