Better error handling when initializing a set view group
Exceptions, errors and exits were not being considered.
Further, it ensures that view group opener processes always
get a reply when such errors happen, instead of hanging
forever waiting for a reply.
When a partition database can't be open, the errors are now
more explicit and helpful as well.

This relates to MB-4626.

Change-Id: Ie2536e4ff80edf9851896799523a7e6f72ddde49
Reviewed-on: http://review.couchbase.org/12600
Tested-by: Farshid Ghods <farshid.ghods@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>