MB-4712 Use couch_util:to_binary in couch_httpd:error_info.
Some of the errors generated by erlang are of the form {Reason,
Stacktrace} where Stacktrace is a list (but obviously not a
string). If such an error is rethrown somewhere in http handler then
couch_httd:error_info will crash since it tries to convert the second
element of the tuple to a binary using list_to_binary function. To
avoid this couch_util:to_binary is now used.

Change-Id: Ic50291fef59f05632b0ed00c6e356f55e5f4048f
Reviewed-on: http://review.couchbase.org/12778
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: buildbot <qe@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>