MB-4697 Unlink couch_view_merger_queue early.
Now all the folder processes are linked to a queue. Hence if some of
the folders are still alive when cleanup code is executed (e.g. if
some meaningful exception was thrown), then the queue will die with
reason 'killed'. But since the queue is unlinked after killing
folders, corresponding EXIT message will be delivered to the main
process. And this EXIT message will be interpreted by
clean_exit_messages as an error. To avoid this we unlink the queue
before killing any of the folder processes.

Change-Id: Ieb237df39056fe9be978cc7778b14a881b86a7c5
Reviewed-on: http://review.couchbase.org/12780
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: buildbot <qe@couchbase.com>
Reviewed-by: Damien Katz <damien@couchbase.com>