Add set view test 09-deletes-cleanup-many-views.t
This test is similar to 08-deletes-cleanup.t with the
exception that verifies the case where the same design
document has multiple views, as this case triggers a
more complex indexer code path.
It also verifies that 2 views with the same map function
(but different reduce functions) share the same btree in
the index file, and that this btree contains the reduce
values for both reduce functions.

Change-Id: I815d215bef9a65dd7ca4f5dc4b75c1c75c68ded1
Reviewed-on: http://review.couchbase.org/12979
Tested-by: buildbot <qe@couchbase.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>