Add set view test 11-updates-cleanup-many-views.t
This test is similar to 10-updates-cleanup.t with the
exception that verifies the case where the same design
document has multiple views, as this case triggers a
more complex indexer code path.
It also verifies that 2 views with the same map function
(but different reduce functions) share the same btree in
the index file, and that this btree contains the reduce
values for both reduce functions.

Change-Id: Iaee868c22cbe9425e5d2074ff17c24773371941e
Reviewed-on: http://review.couchbase.org/12986
Tested-by: buildbot <qe@couchbase.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>