Avoid deadlock when deleting/shutdown database
When deleting a database, couch_server does it synchronously.
The database's process terminate callback in turn does a
synchronous shutdown of the database updater process. However
when this last step is done, it might be the case that the
database updater is calling the database process with the new
database record that results from a compaction that has just
finished successfully. This last call is synchronous and uses
an infinity timeout (because at this time there's much IO going
on as the old database file is being deleted), which makes the
updater wait for the database's main process response forever
(unless it dies) because the latter is in turn waiting for the
updater to process the shutdown message and terminate - this is
a deadlock scenario, which also makes opening or deleting any
other database impossible.

Issue observed in MB-4760.

Change-Id: If31d6609dfe62f52154624f859258cab0293f81e
Reviewed-on: http://review.couchbase.org/13032
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>