Fix indexing of multiple values for the same Key/DocID pair
Some places in the code, like the compactor and cleaner, expected
the first element of each value tuple to be the partition ID. This
was not the case for when the map function emits multiple values
for the same key.
This change makes sure the first element of the value tuple is
always a partition ID.

Issue observed in MB-4760 and MB-4774.

Change-Id: I21187023e3b20fa77530c60caa911bfb3b0bdc1f
Reviewed-on: http://review.couchbase.org/13067
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Damien Katz <damien@couchbase.com>