Making doc_info and body info simpler, for use by couchstore
Simplifying the storage to only store a single body, with content_meta
type to determine if json or raw binary (and why it's not json), and if compressed.

We are now snappy compressing all persisted erlang terms, doc bodies must be
compressed in a separate step.

Removing couch_btree_nif, as we will be using a native updater in ep-engine.

Change-Id: I6822bbc271861748049bca5ebac7ca0f5fe816d4
Reviewed-on: http://review.couchbase.org/13090
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Damien Katz <damien@couchbase.com>