made notify_db_updated consume exits only from parent.MB-4686
When compactor completes it sends exit normal to updater. That message
is usually ignored (see handle_info implementation). But in
notify_db_updated we consumed any exit and rethrowed it. Which is bad
idea. Because we only care about exit message from parent.

That fixes a bunch of crashes that happen after compaction completes.

Damien Katz helped a lot with diagnosing this problem.

Change-Id: I37ac6cbf7a7e4e3cf21eb4a7ffd737d69f87161c
Reviewed-on: http://review.couchbase.org/13138
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>