MB-4811 Unlink all the folders before killing any of them.
Since recently all the folders became linked to a merger queue. Thus
killing any of them causes also queue death. And this death is
propagated to all other folders. And depending on if those other
folders has already been unlinked or not the EXIT message would or
would not be delivered to the main process. And hence when such EXIT
message is delivered to the main process, it's interpreted as an
error.

Change-Id: If69ed31d40613416fcdf3a82df6bf7e5d0dd608a
Reviewed-on: http://review.couchbase.org/13284
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>