Remove unnecessary linking
The compaction fd no longers needs to be linked to the
compaction process because whenever the later is cancelled
or dies, the view group shutdowns the compaction fd
explicitly. Also make sure the view group is not linked to
open files, the ref counters are already responsible for
managing the files.
Added monitoring of the file and ref counter to help debug
MB-4808 as well.

Change-Id: I2764f89b4b6de9b69ebaed4f86b3473bddb24d0a
Reviewed-on: http://review.couchbase.org/13307
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>