Improve set view indexer performance
Instead of having the mapper queue item by item, let
it queue all the items at once. This reduces messaging
with the work queue, reducing the time the mapper gets
blocked waiting the response to gen_server synchronous
calls.
A quick test showed that an index build time decreased
from 4 minutes to 3 minutes and 22 seconds (repeated the
test several times and reached the same results on a
machine with an ssd disk).

Change-Id: I6b2546e51df2ed6c467e584f4188d70eb2449a2b
Reviewed-on: http://review.couchbase.org/13632
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>