Avoid function_clause error for queries with ?debug=true
When merging the results of the local main and replica indexes
and the query has ?debug=true, the merge would fail with a
function_clause error. The solution is just to ignore debug
information in the callback used by the view merger.
We're only interested in the debug information from the view
merging step done at a higher level (multiple nodes, triggered
by ns_server).

Change-Id: I16778154a415b39ce308cbdb5cf7b401da8f66ca
Reviewed-on: http://review.couchbase.org/13677
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>