More efficient index partition state transitions
If a request to make one (or more) partitions active/passive comes
and the partition is still under cleanup, don't block the caller
until the cleanup finishes. Instead reply immediately to caller and
apply the requested state transition later. This new approach
also offers more chances of avoiding unnecessary transitions/IO if
the caller later asks to cleanup the partition that it previously
asked to mark as active/passive and before this later transition
got applied.

This new behaviour, like the old behaviour, also guarantees that if
the caller gets a positive reply, there's no state information loss
if the server crashes after replying to the caller. The pending state
transition is written to the index header and fsync'ed before replying
to the caller.

Change-Id: I0704d66a33856e540e50f2f4f14f623e881acbe7
Reviewed-on: http://review.couchbase.org/14765
Tested-by: buildbot <build@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>