Fix database MVCC snapshot size after compaction
If we have really many documents (and relatively large), and if
compaction ends up writing many of them at file offsets that cause
the documents to be split into 2 or more blocks, the MVCC snapshot
byte size post compaction can exceed the file size. This happens
because it's reusing the old doc info sizes, which don't account for
the new block prefix bytes.

Closes MB-5053.

Change-Id: I41d53b77a9319bbdbd7112144b24fa08d3966cda
Reviewed-on: http://review.couchbase.org/14797
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>