More efficient cleanup resumption
When we're doing a cleanup of a leaf node (kvnode), ignore
any messages to stop the cleanup (and consider them only when
we're at a kpnode/intermediary node). While at a leaf node,
we're not doing any IO, so processing all the KV pairs of a
leaf node is very fast. This ensures that we won't rewrite
a modified version of this node with only some KV pairs removed,
but with all the necessary KV pairs (that belong to partitions
defined by the cleanup bitmask) removed - this guarantees that
when we resume the cleanup we don't visit the leaf node again
(1 disk read, plus 1 disk write). In some cases this can avoid
visiting entire tree branches when we resume the cleanup (there
was only a single leaf node with KV pairs to cleanup).

Closes CBD-63

Change-Id: I574393b89a2afdeb57621d916eb222c0d22cd559
Reviewed-on: http://review.couchbase.org/14799
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>