Set TCP nodelay option when merging views
This has a huge benefical impact on Linux, reducing response
times up to 93% for the most common cases (queries returning
up to 10000 rows) without having a negative impact on the
remaining cases (queries returning thousands or millions of
rows), since we do our own userland buffering.

This relates to CBD-50.

Change-Id: I4b9bbf6bcc8c569028fc144118a6c03c79afb9c1
Reviewed-on: http://review.couchbase.org/15062
Tested-by: buildbot <build@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>