CBD-50 Faster view querying for trivial/common case
If there's only a single set view to merge, trigger a
ligther codepath that doesn't spawn intermediary processes
(priority queue, folder process).
This is a common case in Couchbase Server, a node receives
a view query request and then sends a view merge request to
all other nodes. This request merges (streams) a single view.

This reduces latency by 20% to 30% under non-concurrent and
concurrent scenarios, respectively.

This is part of CBD-50

Change-Id: Id62231b580e3b4805a87b32aec0c47b0005abecb
Reviewed-on: http://review.couchbase.org/15160
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>